+++ server/database.py
@@ -1,0 +1,17 @@
+from schemas import UserInDB
+
+fake_users_db = {
+    "johndoe": {
+        "username": "johndoe",
+        "full_name": "John Doe",
+        "email": "johndoe@example.com",
+        "hashed_password": "$2b$12$EixZaYVK1fsbw1ZfbX3OXePaWxn96p36WQoeG6Lruj3vjPGga31lW",
+        "disabled": False,
+    }
+}
+
+
+def get_user(username: str, db=fake_users_db):
+    if username in db:
+        user_dict = db[username]
+        return UserInDB(**user_dict)
